Commit : Policy

+++ app/Http/Controllers/Api/TodolistController.php
@@ -5,24 +5,48 @@ namespace App\Http\Controllers\Api;
 use Exception;
 use App\Models\Todolist;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Log;
 use App\Http\Controllers\Controller;
 use App\Http\Resources\TodlistResource;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
+
 
 class TodolistController extends Controller
 {
+
+     use AuthorizesRequests;
     /**
      * Display a listing of the resource.
      */
     public function index()
     {
-        $todolist = Todolist::latest()->get();
+
 
         // return response()->json([
         //     'data' => $todolist,
         //     'message' => 'Data is successfully'
         // ],200);
 
-        return TodlistResource::collection($todolist);
+        try {
+            //   $todolist = Todolist::latest()->where->('user_id',auth()->user()->id)->get();
+
+             $todolist = Todolist::where('user_id', auth()->user()->id)
+                            ->latest()
+                            ->get();
+
+               return TodlistResource::collection($todolist);
+        } catch (Exception $error) {
+
+
+            Log::error('Data Failed '. $error->getMessage());
+
+              return response()->json([
+                'message' => 'Data failed to show',
+                'cause'  => $error->getMessage()
+            ],500);
+        }
+
+
     }
 
     /**
@@ -30,6 +54,8 @@ class TodolistController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('create', Todolist::class);
+
         $data = $request->validate([
             'title' => 'required|min:3|max:255',
             'desc' => 'required|min:3|max:255',
@@ -37,6 +63,7 @@ class TodolistController extends Controller
         ]);
 
         try {
+            $data['user_id'] = auth()->user()->id;
             $todolist = Todolist::create($data);
 
             return response()->json([
@@ -56,7 +83,10 @@ class TodolistController extends Controller
      */
     public function show(string $id)
     {
+
         $data = Todolist::find($id);
+
+        $this->authorize('view', $data);
 
         if($data == null){
             return response()->json([
@@ -72,13 +102,16 @@ class TodolistController extends Controller
      */
     public function update(Request $request, string $id)
     {
+        $todolist =  Todolist::find($id);
+
+        $this->authorize('update',$todolist);
+
         $data = $request->validate([
             'title' => 'required|min:3|max:255',
             'desc' => 'required|min:3|max:255',
             'is_done' => 'required|in:1,0'
         ]);
 
-        $todolist =  Todolist::find($id);
 
         if($todolist == null){
             return response()->json([
@@ -113,6 +146,9 @@ class TodolistController extends Controller
                 'message' => 'data is not found'
             ],404);
         }
+
+        $this->authorize('delete', $todolist);
+        
         try {
             $todolist->delete();
